+++ src/main/java/net/minecraft/block/BlockCauldron.java
@@ -33,7 +33,7 @@ public class BlockCauldron extends Block
     public BlockCauldron()
     {
         super(Material.iron, MapColor.stoneColor);
-        this.setDefaultState(this.blockState.getBaseState().withProperty(LEVEL, Integer.valueOf(0)));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(LEVEL, 0));
     }
 
     /**
@@ -81,7 +81,7 @@ public class BlockCauldron extends Block
      */
     public void onEntityCollidedWithBlock(World worldIn, BlockPos pos, IBlockState state, Entity entityIn)
     {
-        int i = state.getValue(LEVEL).intValue();
+        int i = state.getValue(LEVEL);
         float f = (float)pos.getY() + (6.0F + (float)(3 * i)) / 16.0F;
 
         if (!worldIn.isRemote && entityIn.isBurning() && i > 0 && entityIn.getEntityBoundingBox().minY <= (double)f)
@@ -107,7 +107,7 @@ public class BlockCauldron extends Block
             }
             else
             {
-                int i = state.getValue(LEVEL).intValue();
+                int i = state.getValue(LEVEL);
                 Item item = itemstack.getItem();
 
                 if (item == Items.water_bucket)
@@ -218,7 +218,7 @@ public class BlockCauldron extends Block
 
     public void setWaterLevel(World worldIn, BlockPos pos, IBlockState state, int level)
     {
-        worldIn.setBlockState(pos, state.withProperty(LEVEL, Integer.valueOf(MathHelper.clamp_int(level, 0, 3))), 2);
+        worldIn.setBlockState(pos, state.withProperty(LEVEL, MathHelper.clamp_int(level, 0, 3)), 2);
         worldIn.updateComparatorOutputLevel(pos, this);
     }
 
@@ -231,7 +231,7 @@ public class BlockCauldron extends Block
         {
             IBlockState iblockstate = worldIn.getBlockState(pos);
 
-            if (iblockstate.getValue(LEVEL).intValue() < 3)
+            if (iblockstate.getValue(LEVEL) < 3)
             {
                 worldIn.setBlockState(pos, iblockstate.cycleProperty(LEVEL), 2);
             }
@@ -258,7 +258,7 @@ public class BlockCauldron extends Block
 
     public int getComparatorInputOverride(World worldIn, BlockPos pos)
     {
-        return worldIn.getBlockState(pos).getValue(LEVEL).intValue();
+        return worldIn.getBlockState(pos).getValue(LEVEL);
     }
 
     /**
@@ -266,7 +266,7 @@ public class BlockCauldron extends Block
      */
     public IBlockState getStateFromMeta(int meta)
     {
-        return this.getDefaultState().withProperty(LEVEL, Integer.valueOf(meta));
+        return this.getDefaultState().withProperty(LEVEL, meta);
     }
 
     /**
@@ -274,7 +274,7 @@ public class BlockCauldron extends Block
      */
     public int getMetaFromState(IBlockState state)
     {
-        return state.getValue(LEVEL).intValue();
+        return state.getValue(LEVEL);
     }
 
     protected BlockState createBlockState()
