+++ src/main/java/net/minecraft/block/BlockDaylightDetector.java
@@ -31,7 +31,7 @@ public class BlockDaylightDetector extends BlockContainer
     {
         super(Material.wood);
         this.inverted = inverted;
-        this.setDefaultState(this.blockState.getBaseState().withProperty(POWER, Integer.valueOf(0)));
+        this.setDefaultState(this.blockState.getBaseState().withProperty(POWER, 0));
         this.setBlockBounds(0.0F, 0.0F, 0.0F, 1.0F, 0.375F, 1.0F);
         this.setCreativeTab(CreativeTabs.tabRedstone);
         this.setHardness(0.2F);
@@ -46,7 +46,7 @@ public class BlockDaylightDetector extends BlockContainer
 
     public int getWeakPower(IBlockAccess worldIn, BlockPos pos, IBlockState state, EnumFacing side)
     {
-        return state.getValue(POWER).intValue();
+        return state.getValue(POWER);
     }
 
     public void updatePower(World worldIn, BlockPos pos)
@@ -66,9 +66,9 @@ public class BlockDaylightDetector extends BlockContainer
                 i = 15 - i;
             }
 
-            if (iblockstate.getValue(POWER).intValue() != i)
+            if (iblockstate.getValue(POWER) != i)
             {
-                worldIn.setBlockState(pos, iblockstate.withProperty(POWER, Integer.valueOf(i)), 3);
+                worldIn.setBlockState(pos, iblockstate.withProperty(POWER, i), 3);
             }
         }
     }
@@ -158,7 +158,7 @@ public class BlockDaylightDetector extends BlockContainer
      */
     public IBlockState getStateFromMeta(int meta)
     {
-        return this.getDefaultState().withProperty(POWER, Integer.valueOf(meta));
+        return this.getDefaultState().withProperty(POWER, meta);
     }
 
     /**
@@ -166,7 +166,7 @@ public class BlockDaylightDetector extends BlockContainer
      */
     public int getMetaFromState(IBlockState state)
     {
-        return state.getValue(POWER).intValue();
+        return state.getValue(POWER);
     }
 
     protected BlockState createBlockState()
