+++ src/main/java/net/minecraft/enchantment/EnchantmentHelper.java
@@ -82,7 +82,7 @@ public class EnchantmentHelper
             {
                 int j = nbttaglist.getCompoundTagAt(i).getShort("id");
                 int k = nbttaglist.getCompoundTagAt(i).getShort("lvl");
-                map.put(Integer.valueOf(j), Integer.valueOf(k));
+                map.put(j, k);
             }
         }
 
@@ -99,19 +99,19 @@ public class EnchantmentHelper
 
         while (iterator.hasNext())
         {
-            int i = ((Integer)iterator.next()).intValue();
+            int i = (Integer) iterator.next();
             Enchantment enchantment = Enchantment.getEnchantmentById(i);
 
             if (enchantment != null)
             {
                 NBTTagCompound nbttagcompound = new NBTTagCompound();
                 nbttagcompound.setShort("id", (short)i);
-                nbttagcompound.setShort("lvl", (short) enchMap.get(Integer.valueOf(i)).intValue());
+                nbttagcompound.setShort("lvl", (short) enchMap.get(i).intValue());
                 nbttaglist.appendTag(nbttagcompound);
 
                 if (stack.getItem() == Items.enchanted_book)
                 {
-                    Items.enchanted_book.addEnchantment(stack, new EnchantmentData(enchantment, enchMap.get(Integer.valueOf(i)).intValue()));
+                    Items.enchanted_book.addEnchantment(stack, new EnchantmentData(enchantment, enchMap.get(i)));
                 }
             }
         }
@@ -455,7 +455,7 @@ public class EnchantmentHelper
 
                             for (EnchantmentData enchantmentdata1 : list)
                             {
-                                if (!enchantmentdata1.enchantmentobj.canApplyTogether(Enchantment.getEnchantmentById(integer.intValue())))
+                                if (!enchantmentdata1.enchantmentobj.canApplyTogether(Enchantment.getEnchantmentById(integer)))
                                 {
                                     flag = false;
                                     break;
@@ -500,7 +500,7 @@ public class EnchantmentHelper
                             map = Maps.<Integer, EnchantmentData>newHashMap();
                         }
 
-                        map.put(Integer.valueOf(enchantment.effectId), new EnchantmentData(enchantment, i));
+                        map.put(enchantment.effectId, new EnchantmentData(enchantment, i));
                     }
                 }
             }
