+++ src/main/java/net/minecraft/entity/item/EntityArmorStand.java
@@ -76,7 +76,7 @@ public class EntityArmorStand extends EntityLivingBase
     protected void entityInit()
     {
         super.entityInit();
-        this.dataWatcher.addObject(10, Byte.valueOf((byte)0));
+        this.dataWatcher.addObject(10, (byte) 0);
         this.dataWatcher.addObject(11, DEFAULT_HEAD_ROTATION);
         this.dataWatcher.addObject(12, DEFAULT_BODY_ROTATION);
         this.dataWatcher.addObject(13, DEFAULT_LEFTARM_ROTATION);
@@ -818,7 +818,7 @@ public class EntityArmorStand extends EntityLivingBase
             b0 = (byte)(b0 & -2);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean isSmall()
@@ -839,7 +839,7 @@ public class EntityArmorStand extends EntityLivingBase
             b0 = (byte)(b0 & -3);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean hasNoGravity()
@@ -860,7 +860,7 @@ public class EntityArmorStand extends EntityLivingBase
             b0 = (byte)(b0 & -5);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean getShowArms()
@@ -881,7 +881,7 @@ public class EntityArmorStand extends EntityLivingBase
             b0 = (byte)(b0 & -9);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     public boolean hasNoBasePlate()
@@ -905,7 +905,7 @@ public class EntityArmorStand extends EntityLivingBase
             b0 = (byte)(b0 & -17);
         }
 
-        this.dataWatcher.updateObject(10, Byte.valueOf(b0));
+        this.dataWatcher.updateObject(10, b0);
     }
 
     /**
